ui refactoring all functional now
