+++ src/ui/tabs.rs
@@ -14,14 +14,14 @@ pub enum SelectedTab {
 }
 impl SelectedTab {
     /// Get the previous tab, if there is no previous tab return the current tab.
-    fn previous(self) -> Self {
+    pub fn previous(self) -> Self {
         let current_index: usize = self as usize;
         let previous_index = current_index.saturating_sub(1);
         Self::from_repr(previous_index).unwrap_or(self)
     }
 
     /// Get the next tab, if there is no next tab return the current tab.
-    fn next(self) -> Self {
+    pub fn next(self) -> Self {
         let current_index = self as usize;
         let next_index = current_index.saturating_add(1);
         Self::from_repr(next_index).unwrap_or(self)
